feat(components): the settings panel takes its data as props and moves
<SettingsPanel/> is our code; what is AGPL is the data it renders — upstream's labels
and tooltips, its tab tree, its toggle rules. So the component goes to the permissive
package and takes schema, uiTree and toggle as props. Without them it is a working
bare form: every field labelled by its key, one flat page per builder, nothing
disabled. three-slicer/components is the wrapper that passes the full schema, the real
tree and the rule-bound evaluator; a host with its own vocabulary (a translation, its
own labels) passes its own.

three-slicer/toggle's evaluator is ours too, so it becomes makeToggle(rules, schema)
in the permissive package — unbound it disables nothing — and the AGPL side binds
upstream's toggle-rules.json and re-exports the same three functions. Measured:
unbound 0 keys disabled, bound 68 with wall_loops at 0.

Caught by rendering rather than by tests: the default-tree block was placed above the
helper it calls, a TDZ error that took the whole page down and that node never sees
because it does not run JSX.
