+++ packages-mit/src/components/BedShapeEditor.jsx
